+++ src/pyVertexModel/Kg/kgTriAREnergyBarrier.py
@@ -33,7 +33,7 @@ class KgTriAREnergyBarrier(Kg):
                         y2 = Ys[Tris[t].Edge[1], :]
                         y3 = Cell.Faces[f].Centre
 
-                        ys = np.zeros([3, 3, 3], dtype=float)
+                        ys = np.zeros([3, 3, 3], dtype=np.float16)
                         nY = np.zeros([3, 3], dtype=int)
 
                         ys[0, :, :] = [y1, y2, y3]
@@ -60,7 +60,7 @@ class KgTriAREnergyBarrier(Kg):
 
                             w_t[numY] = np.linalg.norm(v_y1) ** 2 - np.linalg.norm(v_y2) ** 2
 
-                            gs = np.zeros(9, dtype=float)
+                            gs = np.zeros(9, dtype=np.float16)
                             gs[0:3] = Set.lambdaR * w_t[numY] * v_y3_1
                             gs[3:6] = Set.lambdaR * w_t[numY] * v_y3_2
                             gs[6:9] = Set.lambdaR * w_t[numY] * v_y3_3
@@ -104,7 +104,7 @@ class KgTriAREnergyBarrier(Kg):
                         y2 = Ys[Tris[t].Edge[1], :]
                         y3 = Cell.Faces[f].Centre
 
-                        ys = np.zeros([3, 3, 3], dtype=float)
+                        ys = np.zeros([3, 3, 3], dtype=np.float16)
                         nY = np.zeros([3, 3], dtype=int)
 
                         ys[0, :, :] = [y1, y2, y3]
@@ -131,7 +131,7 @@ class KgTriAREnergyBarrier(Kg):
 
                             w_t[numY] = np.linalg.norm(v_y1) ** 2 - np.linalg.norm(v_y2) ** 2
 
-                            gs = np.zeros(9, dtype=float)
+                            gs = np.zeros(9, dtype=np.float16)
                             gs[0:3] = Set.lambdaR * w_t[numY] * v_y3_1
                             gs[3:6] = Set.lambdaR * w_t[numY] * v_y3_2
                             gs[6:9] = Set.lambdaR * w_t[numY] * v_y3_3
